requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-generativeai
python-telegram-bot==20.5
python-dotenv
APScheduler
streamlit
requests
